Group ati analytics cypress tests by page type
