+++ package.json
@@ -33,9 +33,12 @@
     "build:test:debug": "rm -rf build && awk '{sub(/LOG_DIR=.+/,\"LOG_DIR='log'\")}1' envConfig/test.env > .env && NODE_ENV=production webpack",
     "cypress:application": "cypress run --browser chrome --spec ./cypress/e2e/application",
     "cypress:pages": "cypress run --browser chrome --spec ./cypress/e2e/pages",
-    "cypress:special-features:ati-analytics": "cypress run --browser chrome --spec ./cypress/e2e/specialFeatures/atiAnalytics",
+    "cypress:ati-analytics:articles": "cypress run --browser chrome --spec ./cypress/e2e/specialFeatures/atiAnalytics/pages/articles",
+    "cypress:ati-analytics:media": "cypress run --browser chrome --spec ./cypress/e2e/specialFeatures/atiAnalytics/pages/media",
+    "cypress:ati-analytics:home-topic-most-read": "cypress run --browser chrome --spec ./cypress/e2e/specialFeatures/atiAnalytics/pages/homeTopicMostRead",
+    "cypress:ati-analytics": "run-p --print-name --aggregate-output --continue-on-error cypress:ati-analytics:articles cypress:ati-analytics:media cypress:ati-analytics:home-topic-most-read",
     "cypress:special-features": "cypress run --browser chrome --spec \"./cypress/e2e/specialFeatures/!(atiAnalytics)/**\"",
-    "cypress": "run-p --print-name --aggregate-output --continue-on-error cypress:application cypress:pages cypress:special-features cypress:special-features:ati-analytics",
+    "cypress": "run-p --print-name --aggregate-output --continue-on-error cypress:application cypress:pages cypress:special-features cypress:ati-analytics",
     "cypress:onefile": "CYPRESS_APP_ENV=live CYPRESS_SMOKE=false cypress run --spec './cypress/e2e/pages/storyPage/*'",
     "cypress:oneservice": "CYPRESS_APP_ENV=live CYPRESS_SMOKE=false CYPRESS_ONLY_SERVICE=mundo cypress run",
     "cypress:oneserviceandonefile": "CYPRESS_APP_ENV=live CYPRESS_SMOKE=false CYPRESS_ONLY_SERVICE=mundo cypress run  --spec './cypress/e2e/pages/storyPage/*'",
